qgis: 3.44.7 -> 4.0.0
Update QGIS to the lastest major version and update packaging based on
upstream packaging located in
https://github.com/qgis/QGIS/tree/release-4_0/nix .

+++ pkgs/applications/gis/qgis/unwrapped.nix
@@ -1,20 +1,18 @@
 {
   lib,
   stdenv,
+
   fetchFromGitHub,
   lndir,
   makeWrapper,
-  mkDerivation,
   replaceVars,
   wrapGAppsHook3,
   wrapQtAppsHook,
 
   withGrass,
   withServer,
-  withWebKit,
 
   darwin,
-  libtasn1,
   bison,
   cmake,
   draco,
@@ -25,6 +23,8 @@
   grass,
   gsl,
   hdf5,
+  libtasn1,
+  libspatialindex,
   libspatialite,
   libzip,
   netcdf,
@@ -35,18 +35,20 @@
   proj,
   protobuf,
   python3,
-  qca-qt5,
+  qca,
   qscintilla,
   qt3d,
+  qt5compat,
   qtbase,
+  qtdeclarative,
   qtkeychain,
   qtlocation,
   qtmultimedia,
   qtsensors,
   qtserialport,
   qtsvg,
-  qtwebkit,
-  qtxmlpatterns,
+  qttools,
+  qtwebengine,
   qwt,
   sqlite,
   txt2tags,
@@ -57,8 +59,7 @@ let
   py = python3.override {
     self = py;
     packageOverrides = self: super: {
-      pyqt5 = super.pyqt5.override {
-        withLocation = true;
+      pyqt6 = super.pyqt6.override {
         withSerialPort = true;
       };
     };
@@ -72,30 +73,30 @@ let
     owslib
     psycopg2
     pygments
-    pyqt5
-    pyqt5-sip
+    pyqt6
     pyqt-builder
     python-dateutil
     pytz
     pyyaml
-    qscintilla-qt5
+    qscintilla-qt6
     requests
     setuptools
     sip
     six
     urllib3
   ];
+
 in
-mkDerivation rec {
-  version = "3.44.7";
+stdenv.mkDerivation rec {
   pname = "qgis-unwrapped";
+  version = "4.0.0";
   outputs = [ "out" ] ++ lib.optional (!stdenv.hostPlatform.isDarwin) "man";
 
   src = fetchFromGitHub {
     owner = "qgis";
     repo = "QGIS";
     rev = "final-${lib.replaceStrings [ "." ] [ "_" ] version}";
-    hash = "sha256-ewYTl6tR1K7Q/AMA3BDlcSaj4mN3yt+1Kfa4a1JSXL0=";
+    hash = "sha256-c1K7R/9OjukBmDmc2jUbf9VKzpf4qbB6OfuKup2wfKA=";
   };
 
   passthru = {
@@ -125,42 +126,47 @@ mkDerivation rec {
     geos
     gsl
     hdf5
+    libpq
+    libspatialindex
     libspatialite
     libzip
     netcdf
     openssl
     pdal
-    libpq
     proj
     protobuf
-    qca-qt5
+    qca
     qscintilla
     qt3d
+    qt5compat
     qtbase
+    qtdeclarative
     qtkeychain
     qtlocation
     qtmultimedia
     qtsensors
     qtserialport
-    qtsvg
-    qtxmlpatterns
+    qttools
+    qtwebengine
     qwt
     sqlite
     txt2tags
     zstd
   ]
   ++ lib.optional withGrass grass
-  ++ lib.optional withWebKit qtwebkit
-  ++ lib.optional stdenv.hostPlatform.isDarwin libtasn1
+  ++ lib.optionals stdenv.hostPlatform.isDarwin [
+    libtasn1
+    qtsvg
+  ]
   ++ pythonBuildInputs;
 
   patches = [
-    (replaceVars ./set-pyqt-package-dirs.patch {
-      pyQt5PackageDir = "${py.pkgs.pyqt5}/${py.pkgs.python.sitePackages}";
-      qsciPackageDir = "${py.pkgs.qscintilla-qt5}/${py.pkgs.python.sitePackages}";
+    (replaceVars ./set-pyqt6-package-dirs.patch {
+      pyQt6PackageDir = "${py.pkgs.pyqt6}/${py.pkgs.python.sitePackages}";
+      qsciPackageDir = "${py.pkgs.qscintilla-qt6}/${py.pkgs.python.sitePackages}";
     })
     (replaceVars ./spatialite-path.patch {
-      spatialiteLib = "${libspatialite}/lib/mod_spatialite.so";
+      spatialiteLib = "${libspatialite}/lib/mod_spatialite${stdenv.hostPlatform.extensions.sharedLibrary}";
     })
   ];
 
@@ -169,6 +175,8 @@ mkDerivation rec {
   env.QT_QPA_PLATFORM_PLUGIN_PATH = "${qtbase}/${qtbase.qtPluginPrefix}/platforms";
 
   cmakeFlags = [
+    "-DWITH_QTWEBENGINE=True"
+
     "-DWITH_3D=True"
     "-DWITH_PDAL=True"
     "-DENABLE_TESTS=False"
@@ -178,10 +186,10 @@ mkDerivation rec {
     "-DWITH_INTERNAL_SPATIALINDEX=True"
   ]
   ++ lib.optionals stdenv.hostPlatform.isDarwin [
-    "-DQGIS_MACAPP_BUNDLE=0" # Don't copy Qt into bundle; we fix paths in postFixup
-    "-DSQLITE3_INCLUDE_DIR=${sqlite.dev}/include" # FindSqlite3.cmake incorrectly assumes framework
+    "-DQGIS_MACAPP_BUNDLE=0"
+    "-DSQLITE3_INCLUDE_DIR=${sqlite.dev}/include"
+    "-DUSE_OPENCL=OFF"
   ]
-  ++ lib.optional (!withWebKit) "-DWITH_QTWEBKIT=OFF"
   ++ lib.optional withServer [
     "-DWITH_SERVER=True"
     "-DQGIS_CGIBIN_SUBDIR=${placeholder "out"}/lib/cgi-bin"
@@ -201,12 +209,11 @@ mkDerivation rec {
   dontWrapGApps = true; # wrapper params passed below
   dontWrapQtApps = stdenv.hostPlatform.isDarwin;
 
+  # GRASS has to be available on the command line even though we baked in the
+  # path at build time using GRASS_PREFIX. Using wrapGAppsHook also prevents
+  # file dialogs from crashing the program on non-NixOS.
   postFixup =
     lib.optionalString (withGrass && stdenv.hostPlatform.isLinux) ''
-      # GRASS has to be available on the command line even though we baked in
-      # the path at build time using GRASS_PREFIX.
-      # Using wrapGAppsHook also prevents file dialogs from crashing the program
-      # on non-NixOS.
       for program in $out/bin/*; do
         wrapProgram $program \
           "''${gappsWrapperArgs[@]}" \
@@ -214,32 +221,46 @@ mkDerivation rec {
       done
     ''
     + lib.optionalString stdenv.hostPlatform.isDarwin ''
-            mkdir -p $out/Applications $out/bin
-            mv $out/QGIS.app $out/Applications/
-            ln -s $out/Applications/QGIS.app/Contents/MacOS/QGIS $out/bin/qgis
+            mkdir -p $out/Applications/QGIS.app $out/bin
+            mv $out/Contents $out/Applications/QGIS.app/
+            ln -s $out/Applications/QGIS.app/Contents/MacOS/qgis $out/bin/qgis
 
             SHORT_VERSION=$(echo "${version}" | cut -d. -f1,2)
             BUNDLE="$out/Applications/QGIS.app"
             FRAMEWORKS="$BUNDLE/Contents/Frameworks"
 
+            for lib in "$FRAMEWORKS"/libqgis*.dylib "$FRAMEWORKS"/libqgispython*.dylib; do
+              [[ -f "$lib" && ! -L "$lib" ]] || continue
+              libname=$(basename "$lib")
+              install_name_tool -id "$FRAMEWORKS/$libname" "$lib" 2>/dev/null || true
+            done
+
             fix_binary() {
               local f="$1"
               [[ -f "$f" ]] || return 0
-              file "$f" | grep -q "Mach-O" || return 0
+              file "$f" | grep -q "Mach-O" || return 0  # spellok
 
               install_name_tool \
-                -change "@loader_path/../lib/libqscintilla2_qt5.dylib" "${qscintilla}/lib/libqscintilla2_qt5.dylib" \
-                -change "@loader_path/../lib/libqt5keychain.dylib" "${qtkeychain}/lib/libqt5keychain.dylib" \
+                -change "$out/lib/libqgis_core.${version}.dylib" "$FRAMEWORKS/libqgis_core.${version}.dylib" \
+                -change "$out/lib/libqgis_gui.${version}.dylib" "$FRAMEWORKS/libqgis_gui.${version}.dylib" \
+                -change "$out/lib/libqgis_analysis.${version}.dylib" "$FRAMEWORKS/libqgis_analysis.${version}.dylib" \
+                -change "$out/lib/libqgis_3d.${version}.dylib" "$FRAMEWORKS/libqgis_3d.${version}.dylib" \
+                -change "$out/lib/libqgis_native.${version}.dylib" "$FRAMEWORKS/libqgis_native.${version}.dylib" \
+                -change "$out/lib/libqgis_app.${version}.dylib" "$FRAMEWORKS/libqgis_app.${version}.dylib" \
+                -change "$out/lib/libqgispython.${version}.dylib" "$FRAMEWORKS/libqgispython.${version}.dylib" \
+                -change "qwt.framework/Versions/6/qwt" "${qwt}/lib/qwt.framework/Versions/6/qwt" \
+                -change "@loader_path/../lib/libqscintilla2_qt6.dylib" "${qscintilla}/lib/libqscintilla2_qt6.dylib" \
+                -change "@loader_path/../lib/libqt6keychain.dylib" "${qtkeychain}/lib/libqt6keychain.dylib" \
                 -change "@loader_path/../lib/libqwt.dylib" "${qwt}/lib/libqwt.dylib" \
-                -change "@loader_path/../../Frameworks/qca-qt5.framework/qca-qt5" "${qca-qt5}/lib/qca-qt5.framework/qca-qt5" \
-                -change "@loader_path/../../../qca-qt5.framework/qca-qt5" "${qca-qt5}/lib/qca-qt5.framework/qca-qt5" \
-                -change "@loader_path/../../../../MacOS/lib/libqscintilla2_qt5.dylib" "${qscintilla}/lib/libqscintilla2_qt5.dylib" \
-                -change "@loader_path/../../../../MacOS/lib/libqt5keychain.dylib" "${qtkeychain}/lib/libqt5keychain.dylib" \
+                -change "@loader_path/../../Frameworks/qca-qt6.framework/qca-qt6" "${qca}/lib/qca-qt6.framework/qca-qt6" \
+                -change "@loader_path/../../../qca-qt6.framework/qca-qt6" "${qca}/lib/qca-qt6.framework/qca-qt6" \
+                -change "@loader_path/../../../../MacOS/lib/libqscintilla2_qt6.dylib" "${qscintilla}/lib/libqscintilla2_qt6.dylib" \
+                -change "@loader_path/../../../../MacOS/lib/libqt6keychain.dylib" "${qtkeychain}/lib/libqt6keychain.dylib" \
                 -change "@loader_path/../../../../MacOS/lib/libqwt.dylib" "${qwt}/lib/libqwt.dylib" \
                 -change "@executable_path/lib/libqwt.dylib" "${qwt}/lib/libqwt.dylib" \
-                -change "@executable_path/lib/libqscintilla2_qt5.dylib" "${qscintilla}/lib/libqscintilla2_qt5.dylib" \
-                -change "@executable_path/lib/libqt5keychain.dylib" "${qtkeychain}/lib/libqt5keychain.dylib" \
-                -change "@executable_path/../Frameworks/qca-qt5.framework/qca-qt5" "${qca-qt5}/lib/qca-qt5.framework/qca-qt5" \
+                -change "@executable_path/lib/libqscintilla2_qt6.dylib" "${qscintilla}/lib/libqscintilla2_qt6.dylib" \
+                -change "@executable_path/lib/libqt6keychain.dylib" "${qtkeychain}/lib/libqt6keychain.dylib" \
+                -change "@executable_path/../Frameworks/qca-qt6.framework/qca-qt6" "${qca}/lib/qca-qt6.framework/qca-qt6" \
                 -change "@loader_path/../../../qgis_core.framework/qgis_core" "$FRAMEWORKS/qgis_core.framework/Versions/$SHORT_VERSION/qgis_core" \
                 -change "@loader_path/../../../qgis_gui.framework/qgis_gui" "$FRAMEWORKS/qgis_gui.framework/Versions/$SHORT_VERSION/qgis_gui" \
                 -change "@loader_path/../../../qgis_analysis.framework/qgis_analysis" "$FRAMEWORKS/qgis_analysis.framework/Versions/$SHORT_VERSION/qgis_analysis" \
@@ -248,14 +269,21 @@ mkDerivation rec {
                 "$f" 2>/dev/null || true
             }
 
-            fix_binary "$BUNDLE/Contents/MacOS/QGIS"
-            for lib in "$BUNDLE/Contents/MacOS/lib"/*.dylib; do fix_binary "$lib"; done
+            fix_binary "$BUNDLE/Contents/MacOS/qgis"
+            for bin in "$BUNDLE/Contents/MacOS"/*; do fix_binary "$bin"; done
+            for lib in "$FRAMEWORKS"/*.dylib; do fix_binary "$lib"; done
+            if [[ -d "$BUNDLE/Contents/MacOS/lib" ]]; then
+              for lib in "$BUNDLE/Contents/MacOS/lib"/*.dylib; do fix_binary "$lib"; done
+            fi
             for fw in qgis_core qgis_gui qgis_analysis qgis_3d qgis_native; do
               fix_binary "$FRAMEWORKS/$fw.framework/Versions/$SHORT_VERSION/$fw"
               [[ -f "$FRAMEWORKS/$fw.framework/$fw" && ! -L "$FRAMEWORKS/$fw.framework/$fw" ]] && \
                 fix_binary "$FRAMEWORKS/$fw.framework/$fw"
             done
             for plugin in "$BUNDLE/Contents/PlugIns/qgis"/*.so; do fix_binary "$plugin"; done
+
+            # Fix Python binding .so files in Frameworks/qgis
+            for pyso in "$BUNDLE/Contents/Frameworks/qgis"/*.so; do fix_binary "$pyso"; done
 
             ${lib.optionalString withGrass ''
               fix_binary "$FRAMEWORKS/qgisgrass8.framework/Versions/$SHORT_VERSION/qgisgrass8"
@@ -273,22 +301,21 @@ mkDerivation rec {
               fix_binary "$BUNDLE/Contents/MacOS/lib/libqgis_server.${version}.dylib"
             ''}
 
-            # Merge Python packages (lndir handles namespace packages correctly)
             QGIS_PYTHON="$BUNDLE/Contents/Resources/python"
-            for pkg in ${
-              lib.concatMapStringsSep " " (p: "${p}/${py.pkgs.python.sitePackages}") (
-                py.pkgs.requiredPythonModules pythonBuildInputs
-              )
-            }; do
-              [[ -d "$pkg" ]] && lndir -silent "$pkg" "$QGIS_PYTHON"
-            done
-            # Remove broken symlinks
-            find "$QGIS_PYTHON" -type l ! -exec test -e {} \; -delete
+            if [[ -d "$QGIS_PYTHON" ]]; then
+              for pkg in ${
+                lib.concatMapStringsSep " " (p: "${p}/${py.pkgs.python.sitePackages}") (
+                  py.pkgs.requiredPythonModules pythonBuildInputs
+                )
+              }; do
+                [[ -d "$pkg" ]] && lndir -silent "$pkg" "$QGIS_PYTHON"
+              done
+              find "$QGIS_PYTHON" -type l ! -exec test -e {} \; -delete
+            fi
 
-            # Create merged Qt plugins directory in the bundle (LSEnvironment is unreliable)
             mkdir -p "$BUNDLE/Contents/PlugIns"
             lndir -silent "${qtbase}/${qtbase.qtPluginPrefix}" "$BUNDLE/Contents/PlugIns"
-            lndir -silent "${qtsvg.bin}/${qtbase.qtPluginPrefix}" "$BUNDLE/Contents/PlugIns"
+            lndir -silent "${qtsvg}/${qtbase.qtPluginPrefix}" "$BUNDLE/Contents/PlugIns"
 
             cat > "$BUNDLE/Contents/Resources/qt.conf" << 'EOF'
       [Paths]
